+++ src/components/App/App.js
@@ -1,5 +1,6 @@
 import React, { useState } from "react";
 import { HashRouter as Router, Switch, Route } from "react-router-dom";
+import { Up, scrollToAnchor } from "../Anchor";
 import Navbar from "../Navbar";
 import AdminNavbar from "../AdminNavbar";
 import Footer from "../Footer";
@@ -7,6 +8,7 @@ import {
   HomePage,
   AboutPage,
   LoginPage,
+  LogoutPage,
   RegisterPage,
   ProductListPage,
   ProductPage,
@@ -22,6 +24,7 @@ import {
   AdminMemberPage,
   AdminOrderListPage,
   AdminNewsPage,
+  AdminCategoryPage,
 } from "../../pages";
 
 function App() {
@@ -30,6 +33,9 @@ function App() {
   return (
     <Router>
       {user === "customer" ? <Navbar /> : <AdminNavbar />}
+      <Up onClick={() => scrollToAnchor("top")}>
+        <span>⇧</span>
+      </Up>
       <Switch>
         <Route exact path="/">
           <HomePage />
@@ -39,6 +45,9 @@ function App() {
         </Route>
         <Route exact path="/login">
           <LoginPage />
+        </Route>
+        <Route exact path="/logout">
+          <LogoutPage />
         </Route>
         <Route exact path="/register">
           <RegisterPage />
@@ -73,8 +82,11 @@ function App() {
         <Route exact path="/admin/products">
           <AdminProductListPage />
         </Route>
-        <Route exact path="/admin/product/:id">
+        <Route path="/admin/product/">
           <AdminProductPage />
+        </Route>
+        <Route path="/admin/category/">
+          <AdminCategoryPage />
         </Route>
         <Route exact path="/admin/member">
           <AdminMemberPage />
