+++ src/pages/admin/AdminCategory/AdminCategory.js
@@ -1,0 +1,3 @@
+export default function AdminCategory() {
+  return <div>CATEGORY_</div>;
+}
