+++ src/pages/user/LogoutPage/LogoutPage.js
@@ -1,0 +1,3 @@
+export default function LogoutPage() {
+  return <div>LOGOUT</div>;
+}
